requirements.txt
apache-beam==2.49.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastavro==1.8.2
fasteners==0.18
google-api-core==2.11.1
google-apitools==0.5.31
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.31.1
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.21.0
google-cloud-core==2.3.3
google-cloud-datastore==2.17.0
google-cloud-dlp==3.12.2
google-cloud-language==2.11.0
google-cloud-pubsub==2.18.3
google-cloud-pubsublite==1.8.3
google-cloud-recommendations-ai==0.10.4
google-cloud-resource-manager==1.10.3
google-cloud-spanner==3.40.1
google-cloud-storage==2.10.0
google-cloud-videointelligence==2.11.3
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
hdfs==2.7.2
httplib2==0.22.0
idna==3.4
numpy==1.24.4
oauth2client==4.1.3
objsize==0.6.1
orjson==3.9.5
overrides==6.5.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
rsa==4.9
ruff==0.0.285
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
zstandard==0.21.0
